further breaking down of the DefaultService. post becomes post + add
